fix: liberate double types

+++ Source/rclUE/Public/Msgs/ROS2Float64MultiArray.h
@@ -27,13 +27,13 @@ public:
     // rclc type: example_interfaces__msg__MultiArrayLayout layout
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite)
-    TArray<float> data;
+    TArray<double> data;
     // rclc type: rosidl_runtime_c__double__Sequence data
 
     void SetFromROS2(const ros_msg_c_typename& in_ros_data)
     {
         layout.SetFromROS2(in_ros_data.layout);
-        data = ROS2MsgToUE::FromSequence<float>(in_ros_data.data);
+        data = ROS2MsgToUE::FromSequence<double>(in_ros_data.data);
     }
 
     void SetROS2(ros_msg_c_typename& out_ros_data) const
