+++ Source/rclUE/Public/Msgs/ROS2UInt32.h
@@ -21,7 +21,7 @@ struct RCLUE_API FROSUInt32
 public:
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite)
-    int data;
+    int64 data;
     // rclc type: uint32_t data
 
     void SetFromROS2(const ros_msg_c_typename& in_ros_data)
